Cache MethodToCallFromOutside and create blueprints in parallel
The function getMethodToCallForDependency is called on every dependency.
This function creates a new ModuleBluePrint to extract
methodToCallFromOutside. In a project with N modules there can be upto
N*(N-1)/2 dependencies, meaning that the same number of ModuleBluePrint will
be generated and deleted.

This change adds a cache to store the values of methodToCallFromOutside,
needing now only N ModuleBluePrint.

Also, now the constructor of ProjectBlueprint create moduleBlueprints in
parallel.

A project with 1000 modules wnet from taking around 100 seconds to just
18 (5x imrpovement in time).