Save dependencies adjacency matrix in a png image
For large projects with many module dependencies, the dependency graph
can be dificult to visualize. Now, a png image is generated with a
representation of the adjacency matrix.

The first row and first columns are headers, the colors there represent
the type of module:
  - Blue: Android app
  - Green: Android module
  - Yellow: Java module

If the cell in row r and column c is green, it means that the module
associated with that row has a dependency with the module associated
with that column. We can add different colors for different dependency
methods later (api, implementation, testImplementation, etc).